refactor: 인터페이스와 구현체명 변경
