+++ src/main/kotlin/model/NumberGenerator.kt
@@ -1,0 +1,5 @@
+package model
+
+interface NumberGenerator {
+    fun generate(): Int
+}
